+++ src/predictor_pu/predictor_pu.c
@@ -348,98 +348,7 @@ static bam_header_t *bam_header_dup(const bam_header_t *h0)
 	}
 	return h;
 }
-static void append_header_text(bam_header_t *header, char* text, int len)
-{
-	int x = header->l_text + 1;
-	int y = header->l_text + len + 1; // 1 byte null
-	if (text == 0) return;
-#if 0
-	/* this is only sensible if l_text was rounded up when text was first allocated
-           We call append_header_text() with a header generated by bam_header_dup()
-           which doesn't round up l_text so we should not round up */
-        kroundup32(x); 
-	kroundup32(y);
-#endif
-	if (x < y) header->text = (char*)realloc(header->text, y);
-	strncpy(header->text + header->l_text, text, len); // we cannot use strcpy() here.
-	header->l_text += len;
-	header->text[header->l_text] = 0;
-}
 
-static void bam_header_add_pg(Settings *s, bam_header_t *bam_header) {
-    char *id = "pb_cal";
-    char *pn = "predictor_pu";
-    char *pp = NULL;
-    char *ds = "A program to apply a calibration table";
-    char *id2, *pg;
-    int id2size, id2num, pgsize, pglen;
-
-    if (NULL == bam_header) {
-        fprintf(stderr,"ERROR: No bam header\n");
-        exit(EXIT_FAILURE);
-    }
-
-    if (NULL == bam_header->text) {
-        fprintf(stderr,"ERROR: No text in bam header\n");
-        exit(EXIT_FAILURE);
-    }
-
-    id2size = 10 + strlen(id);
-    id2 = smalloc(id2size);
-    id2num = 0;
-    while(id2num < 10) {
-        char *text = strdup(bam_header->text);
-        char *hl, *endl, *endt;
-        sprintf(id2, (id2num > 0 ? "%s%d" : "%s"), id, id2num);
-        hl = text;
-        while (0 < strlen(hl)) {
-            if (NULL == (endl = strchr(hl, '\n'))) {
-                die("ERROR: Corrupt bam header \"%s\"\n", hl);
-            }
-            *endl = 0;
-            if (0 == memcmp(hl, "@PG", 3)) {
-                if (NULL == (pp = strstr(hl, "ID:"))) {
-                    die("ERROR: No ID in PG line \"%s\"\n", hl);
-                }
-                pp += 3;
-                if (NULL != (endt = strchr(pp, '\t'))) {
-                    *endt = 0;
-                }
-                if (0 == strcmp(pp, id2)) {
-                    /* an existing ID matches the new ID */
-                    break;
-                }
-            }
-            hl = endl + 1;
-        }
-        if( 0 == strlen(hl)) {
-            /* the new ID doesn't match an existing ID */
-            free(text);
-            break;
-        }
-        /* the new ID matches an existing ID, increment the number and try again */
-        id2num++;
-        free(text);
-    }
-    if (id2num == 10) {
-        die("ERROR: Header already contains PG lines with ID=%s .. ID=%s\n", id, id2);
-    }
-
-    pgsize = 128 + strlen(pn) + strlen(pn) + strlen(ds) + strlen(version) + strlen(s->cmdline);
-    if( NULL != pp ){
-        pgsize += strlen(pp);
-        pg = smalloc(pgsize);
-        pglen = snprintf(pg, pgsize, "@PG\tID:%s\tPN:%s\tPP:%s\tDS:%s\tVN:%s\tCL:%s\n", id2, pn, pp, ds, version, s->cmdline);
-    }else{
-        pg = smalloc(pgsize);
-        pglen = snprintf(pg, pgsize, "@PG\tID:%s\tPN:%s\tDS:%s\tVN:%s\tCL:%s\n", id2, pn, ds, version, s->cmdline);
-    }
-    assert(pglen < pgsize);
-    
-    append_header_text(bam_header, pg, pglen);
-
-    free(pg);
-}
 
 /*
   Calculate purity, update quality value and write bam file
@@ -977,7 +886,7 @@ int main(int argc, char **argv) {
         strcat(out_mode, "u");
 
     out_bam_header = bam_header_dup(fp_input_bam->header);
-    bam_header_add_pg(&settings, out_bam_header);
+    bam_header_add_pg("pb_cal", "predictor_pu", "A program to apply a calibration table", settings.cmdline, out_bam_header);
 
     out_bam_file = (NULL == settings.output ? aprintf("/dev/stdout") : aprintf("%s/%s", settings.working_dir, settings.output));
     fp_output_bam = samopen(out_bam_file, out_mode, out_bam_header);
